feat: prepare DB data, extract and comment init db method
